Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

Only thing is if you press the power button in Charge Only Mode it'll mess up and show the homescreen and have the colors on the screen change around. However it goes back to the battery icon and no functionality is lost. I am unsure if this issue is specific to the Droid2, or if it affects the shadow and droid2we as well.

Patch-set: 3